http: support http endpoints outside the agent package
Refactor how endpoints are registerd to allow endpoints from other
packages.

This is a small incremental step toward a number of objectives:

  1. decreasing the size of the `agent` package by allowing things to be
     extracted into other packages.
  2. reducing the size of the `Agent` struct by allowing endpoints to
     accept a smaller interface, which may not be `Agent`
  3. it may also help support the v2 API effort by allowing for
     endpoints with a different interface.

Generally, avoiding registration in `init()` should make code easier to
work with, and easier to test.

Also create the httpHandlers in Agent.New so that in the future we can easily
pass in the handlers and apiServer, and so that we can set config on it
without needing an Agent field.