Sync node info when updating sync state
Previously we had a variable that would track whether the node's
information was in sync with the servers.

This was a shared variable in the local state which required locking to
read/write.

Rather than guarding access to this variable we now sync node updates
when we see that the node info is out of sync. This only happens in one
location: updateSyncState().

Additionally, rather than SkippingNodeUpdate based on the shared
variable we now let the servers determine whether the node update should
be skipped. This is checked in structs.RegisterRequest.ChangesNode().