agent/checks: improve local dev feedback loop
Add t.Skip for slowest tests

Remove some t.Parallel for tests that are already pretty fast

Always use port 127.0.0.1:0 so that the OS picks a random port.
Previously the hardcoded port could have conflicted with some other
package.

Start a separate gRPC server for each test case that uses it. This
ensures that if no gRPC tests are run, we never start a GRPC server. It
only seems to add at most 10ms to the test run time, and removes the
need for a TestMain, so it seems like a reasonable trade-off.