testing: make a separate module for test/integration
By using a separate module for test/integration we can isolate
dependencies we use in these tests from the production dependencies of
Consul. Isolating dependencies between these two modules makes it easier
for us to manage and update them independently.

We have had some problems with nested go modules in the past, but I
believe all of those problems were a result of moving a package path
from one module to another.

In this case we are explicitly using a module path different from the
previous package path, so we should not have that problem.