gossip: refactor setupSerf
Extract out some functions which can be re-used in some cases.
Start a clientGossip struct which will eventually be able to implement the delegate
membership methods.

This allows us to share more code between client and server.
It is also a small step toward moving dependenct setup out of the constructors
for Client and Server.